+++ experimental/CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## Next
 
-PerformanceTimelineInstrumentation: add beforeEmit hook to make it possible to mutate performance
-entries.
+### PerformanceTimelineInstrumentation
+
+- Add: beforeEmit hook to make it possible to mutate performance entries.
+
+#### Breaking Changes
+
+- Remove: `skipEntries` entries are no removed in favor of the `beforeEmit` hook. You can find an
+  [example how to skip entries in the README.](https://github.com/grafana/faro-web-sdk/blob/a83d2e56b7289ea81a1d0f87c03f73d04bd44e38/experimental/instrumentation-performance-timeline/README.md#example-skip-backforward-navigation-and-page-reloads-to-remove-non-human-visible-navigation)
 
 ## 1.0.5